grab combobox data when submitting

+++ platforms/ios/www/js/index.js
@@ -67,6 +67,8 @@ function LaunchKiosk()
     }
     else
     {
+	   if ( !isOfflineMode() )
+	   {
         localStorage.clear();
         $.ajax({
                type: "GET",
@@ -80,46 +82,8 @@ function LaunchKiosk()
                 var jsonobj = JSON.parse(data);
                 if(jsonobj.Status == 200)
                    {
-                   
-                        localStorage.setItem("key", $('#txtAccessKey').val());
-                        localStorage.setItem("deptname", jsonobj.Data.DeptName);
-                        localStorage.setItem("appdescription", jsonobj.Data.AppDescription);
-                        localStorage.setItem("appid", jsonobj.Data.AppId);
-                        localStorage.setItem("rsvp", jsonobj.Data.IsRSVPOnly);
-                        localStorage.setItem("anon", jsonobj.Data.IsAnonymous);
-                        localStorage.setItem("enforcedeligibility", jsonobj.Data.IsEnforcedEligibility);
-                        localStorage.setItem("initialintakestatus", jsonobj.Data.InitialIntakeStatus);
-                        //localStorage.setItem("hasreasons", jsonobj.Data.HasReasons);
-                        localStorage.setItem("questions", JSON.stringify(jsonobj.Data.Questions));
-                       localStorage.setItem("allowregistration", jsonobj.Data.AllowRegistration);
-                       
-                       localStorage.setItem("eligibilitytype", jsonobj.Data.EligibilityType);
-                       
-                       localStorage.setItem("locations", JSON.stringify(jsonobj.Data.Locations));
-                       var location_array = jsonobj.Data.Locations;
-                       if(location_array.length < 2)
-                       {
-                        if(location_array.length == 1)
-                        {
-                            localStorage.setItem("selLocationID", location_array[0].ID);
-                        }
-                        window.open("login.html", "_self");
-                       }
-                        else
-                        {
-                            $('#div-locations').removeClass('invisible');
-                            //$('#div-locations').empty();
-                           for(var i = 0 ; i < location_array.length ; i++)
-                           {
-                                $('#ddl-locations').append($('<option></option>').val(location_array[i].ID).html(location_array[i].Name));
-                           }
-                        }
-               
-
-                       
-                        //window.open("login.html?key=" + $('#txtAccessKey').val() + "&deptname=" + jsonobj.Data.DeptName + "&appid=" + jsonobj.Data.AppId + "&rsvp=" + jsonobj.Data.IsRSVPOnly + "&anon=" + jsonobj.Data.IsAnonymous + "&initialintakestatus=" + jsonobj.Data.InitialIntakeStatus,"_self");
-                        //window.open("login.html", "_self");
-                    }
+				     RedirectToLoginPage(jsonobj);
+				   }
                 else
                 {
                     showDialog("The access key you entered is incorrect. Please reenter your access key.");
@@ -133,8 +97,59 @@ function LaunchKiosk()
                 $('body').removeClass('ajax-spinner');
                }
         });
+		}
+		else{
+		  //get offline data
+		  var offlineData = GetOfflineData();
+		  var jsonobj = JSON.parse(offlineData);
+		  RedirectToLoginPage(jsonobj);
+		}
     }
     
+}
+
+function RedirectToLoginPage(jsonobj)
+{                  
+		localStorage.setItem("key", $('#txtAccessKey').val());
+		localStorage.setItem("deptname", jsonobj.Data.DeptName);
+		localStorage.setItem("appdescription", jsonobj.Data.AppDescription);
+		localStorage.setItem("appid", jsonobj.Data.AppId);
+		localStorage.setItem("rsvp", jsonobj.Data.IsRSVPOnly);
+		localStorage.setItem("anon", jsonobj.Data.IsAnonymous);
+		localStorage.setItem("enforcedeligibility", jsonobj.Data.IsEnforcedEligibility);
+		localStorage.setItem("initialintakestatus", jsonobj.Data.InitialIntakeStatus);
+		//localStorage.setItem("hasreasons", jsonobj.Data.HasReasons);
+		localStorage.setItem("questions", JSON.stringify(jsonobj.Data.Questions));
+	   localStorage.setItem("allowregistration", jsonobj.Data.AllowRegistration);
+	   
+	   localStorage.setItem("eligibilitytype", jsonobj.Data.EligibilityType);
+	   
+	   localStorage.setItem("locations", JSON.stringify(jsonobj.Data.Locations));
+	   
+	   var location_array = jsonobj.Data.Locations;
+	   if(location_array.length < 2)
+	   {
+		if(location_array.length == 1)
+		{
+			localStorage.setItem("selLocationID", location_array[0].ID);
+		}
+		window.open("login.html", "_self");
+	   }
+		else
+		{
+			$('#div-locations').removeClass('invisible');
+			//$('#div-locations').empty();
+		   for(var i = 0 ; i < location_array.length ; i++)
+		   {
+				$('#ddl-locations').append($('<option></option>').val(location_array[i].ID).html(location_array[i].Name));
+		   }
+		}
+
+
+	   
+		//window.open("login.html?key=" + $('#txtAccessKey').val() + "&deptname=" + jsonobj.Data.DeptName + "&appid=" + jsonobj.Data.AppId + "&rsvp=" + jsonobj.Data.IsRSVPOnly + "&anon=" + jsonobj.Data.IsAnonymous + "&initialintakestatus=" + jsonobj.Data.InitialIntakeStatus,"_self");
+		//window.open("login.html", "_self");
+                    
 }
 
 function SetTestData()
